% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-names.R
\docType{data}
\name{common_names}
\alias{common_names}
\title{Common US names}
\format{
A data frame with 10 rows and 2 columns
\describe{
\item{name}{The name}
\item{n}{The total number of entries in the database}
}
}
\usage{
common_names
}
\description{
The then most common used names in the US according to the
babynames package.
}
\keyword{datasets}
