% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello_world}
\alias{hello_world}
\title{Says hello}
\usage{
hello_world(name, emotion = c("excited", "questioning", "disbelief"))
}
\arguments{
\item{name}{string with the name of the person to greet.}

\item{emotion}{character with the emotion.}
}
\value{
The input name, invisibly |>
}
\description{
Greets you personally and with emotion
}
\details{
This will print to the console. The available emotions might depend on the
supported emotions in your machine.
}
\examples{
hello_world("CRAN", emotion = "disbelief")

}
